backend/requirements.txt
numpy<2.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic==2.7.1
pydantic-settings==2.2.1
langgraph==0.1.19
anthropic==0.26.0
httpx==0.27.0
chromadb==0.5.23
openai==1.30.1
sentence-transformers==3.0.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-multipart==0.0.9
pillow==10.3.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
langfuse==2.21.3
aiosmtplib==3.0.1
python-jose==3.3.0
pytest==8.2.0
pytest-asyncio==0.23.7
anyio==4.4.0
chromadb==0.5.23
numpy<2.0.0
